# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Order Management API",
        "description": "This API handles the management of customer orders",
        "contact": {
            "email": "aloshmohammad2001@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/change-email": {
            "post": {
                "tags": [
                    "Account Sittings"
                ],
                "summary": "Change user email",
                "description": "Update the email address of the currently authenticated user.",
                "operationId": "25a7f0535eee87440aa64acf85feb0c2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "Email"
                                ],
                                "properties": {
                                    "Email": {
                                        "description": "New email address for the user.",
                                        "type": "string",
                                        "format": "email",
                                        "example": "newemail@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Email successfully updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Email successfully updated"
                                        },
                                        "data": {
                                            "properties": {
                                                "Id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "Email": {
                                                    "type": "string",
                                                    "example": "newemail@example.com"
                                                },
                                                "UserName": {
                                                    "type": "string",
                                                    "example": "newemail@example.com"
                                                },
                                                "PhoneNumber": {
                                                    "type": "string",
                                                    "example": "1234567890"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "Email": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The Email has already been taken."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/change-password": {
            "post": {
                "tags": [
                    "Account Sittings"
                ],
                "summary": "Change user password",
                "description": "Allow the authenticated user to change their current password.",
                "operationId": "8182e64143ac402312277ee28a7c9fc0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "current_password",
                                    "new_password"
                                ],
                                "properties": {
                                    "current_password": {
                                        "description": "Current password of the user.",
                                        "type": "string",
                                        "format": "password",
                                        "example": "currentPassword123"
                                    },
                                    "new_password": {
                                        "description": "New password for the user. Must be at least 6 characters long.",
                                        "type": "string",
                                        "format": "password",
                                        "example": "newPassword123"
                                    },
                                    "new_password_confirmation": {
                                        "description": "Confirmation of the new password.",
                                        "type": "string",
                                        "format": "password",
                                        "example": "newPassword123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password successfully updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Password successfully updated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Current password is incorrect",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Current password is incorrect"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "current_password": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The current password field is required."
                                                    }
                                                },
                                                "new_password": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The new password field is required."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/add-phone-number": {
            "post": {
                "tags": [
                    "Account Sittings"
                ],
                "summary": "Add phone number to user profile",
                "description": "Allow the authenticated user to add or update their phone number.",
                "operationId": "d8ce9351b72968752de96e34a5d638f3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "PhoneNumber"
                                ],
                                "properties": {
                                    "PhoneNumber": {
                                        "description": "The phone number to be added or updated.",
                                        "type": "string",
                                        "example": "123-456-7890"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Phone number successfully added",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Phone number successfully added"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "PhoneNumber": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The PhoneNumber field is required."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}